fix: resolve builds error
